Automate the creation of the toc for the docs

+++ md.sh
@@ -2,6 +2,8 @@
 #
 cd ${0%\/*}
 
+[ $# -ne 1 ] && exit 1
+
 cp "UNI/Practica1/${1}" .
 rm -fr UNI
 sed -i s'/..\/..\/_resources/_resources/g' "${1}"
@@ -9,7 +11,7 @@ sed -i s'/..\/..\/_resources/_resources/g' "${1}"
 #get rid of the jopin export header
 sed -i '1,8d' "${1}"
 #generate the toc
-pandoc -s --toc "${1}" -o out.md
+pandoc -s --toc --toc-depth=5 "${1}" -o out.md
 
 mv out.md "${1}"
 #get rid of #{balalalla} stuff on header
